Simulation: Fill in gaps for RFC 5228.
This fixes issue #78.

* "stop" without a prior action triggers implicit keep
* Trim trailing whitespace.
* Fixes :localpart whit a bare address to match spec
* Adds backslash escaping for glob matches
* Update scanner & parser to interpret siE quantifier into the resulting
  numeric.

+++ src/AST/ASTNumeric.cc
@@ -9,11 +9,11 @@ ASTNumeric::ASTNumeric( yy::location location )
 { 
 }
 
-ASTNumeric::ASTNumeric( int number)
+ASTNumeric::ASTNumeric( long number)
     : _number( number )
 {
 }
-ASTNumeric::ASTNumeric( yy::location location, int number)
+ASTNumeric::ASTNumeric( yy::location location, long number)
     : ASTNode( location )
     , _number( number )
 {
